+++ src/main/java/com/jxm/yiti/resp/WxQuestionQueryResp.java
@@ -15,7 +15,7 @@ public class WxQuestionQueryResp {
 
     private Integer importanceLevel;
 
-    private Integer answerId;
+    private Long answerId;
 
     public String getTitle() {
         return title;
@@ -69,11 +69,11 @@ public class WxQuestionQueryResp {
         return importanceLevel;
     }
 
-    public Integer getAnswerId() {
+    public Long getAnswerId() {
         return answerId;
     }
 
-    public void setAnswerId(Integer answerId) {
+    public void setAnswerId(Long answerId) {
         this.answerId = answerId;
     }
